+++ cosinesimilarity.h
@@ -1,3 +1,3 @@
 extern CvScalar cosine_similarity(CvMat query, CvMat patch);
-extern int Resemblance(CvScalar map);
+extern double Resemblance(CvScalar map);
